Fixed the hover for about page
