+++ src/data/officer-data.js
@@ -1,0 +1,52 @@
+const officers = [
+    {
+        name: "Vivian Lee",
+        image: "https://content.fortune.com/wp-content/uploads/2019/01/boo.jpg",
+        position: "President",
+        link: "https://viviaann.github.io/",
+        fun_fact: "",
+        email: "vyl003@ucsd.edu",
+        github: "https://github.com/Viviaann"
+    },
+    {
+        name: "Chau Vu",
+        image:
+            "https://i.pinimg.com/originals/8a/7b/5e/8a7b5e0d1b237ca10e006f0c986b5fe6.jpg",
+        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
+        position: "VP",
+        fun_fact: '"i\'m a fun-sized, super nice guy"',
+        email: "chv004@ucsd.edu",
+        github: "https://github.com/cqvu"
+    },
+    {
+        name: "Brendan Lau",
+        image:
+            "https://hips.hearstapps.com/ghk.h-cdn.co/assets/17/30/2560x1280/landscape-1500925839-golden-retriever-puppy.jpg?resize=480:*",
+        position: "Project Chair",
+        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
+        fun_fact: '"it\'s everyday bro, do you give up at night?"',
+        email: "bwlau@ucsd.edu",
+        github: "https://github.com/blau0123"
+    },
+    {
+        name: "Brandon Foey",
+        image:
+            "https://thespinoff.co.nz/wp-content/uploads/2019/03/GettyImages-139720130.jpg",
+        position: "Workshop Chair",
+        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
+        fun_fact: "",
+        email: "bfoey@ucsd.edu",
+        github: "https://github.com/BrandonFoey"
+    },
+    {
+        name: "Bao Hoang",
+        image: "https://cdn1-www.dogtime.com/assets/uploads/2019/04/DogSilly.jpg",
+        position: "Event Chair",
+        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
+        fun_fact: '"I have a good personality"',
+        email: "bghoang@ucsd.edu",
+        github: "https://github.com/bghoang"
+    }
+];
+
+export default officers;
