+++ src/routes/about.js
@@ -7,56 +7,9 @@ import EmailIcon from "@material-ui/icons/Email";
 import GitHubIcon from "@material-ui/icons/GitHub";
 import Footer from '../components/footer'
 
-const officers = [
-    {
-        name: "Vivian Lee",
-        image: "https://content.fortune.com/wp-content/uploads/2019/01/boo.jpg",
-        position: "President",
-        link: "https://viviaann.github.io/",
-        fun_fact: "",
-        email: "vyl003@ucsd.edu",
-        github: "https://github.com/Viviaann"
-    },
-    {
-        name: "Chau Vu",
-        image:
-            "https://i.pinimg.com/originals/8a/7b/5e/8a7b5e0d1b237ca10e006f0c986b5fe6.jpg",
-        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
-        position: "VP",
-        fun_fact: '"i\'m a fun-sized, super nice guy"',
-        email: "chv004@ucsd.edu",
-        github: "https://github.com/cqvu"
-    },
-    {
-        name: "Brendan Lau",
-        image:
-            "https://hips.hearstapps.com/ghk.h-cdn.co/assets/17/30/2560x1280/landscape-1500925839-golden-retriever-puppy.jpg?resize=480:*",
-        position: "Project Chair",
-        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
-        fun_fact: '"it\'s everyday bro, do you give up at night?"',
-        email: "bwlau@ucsd.edu",
-        github: "https://github.com/blau0123"
-    },
-    {
-        name: "Brandon Foey",
-        image:
-            "https://thespinoff.co.nz/wp-content/uploads/2019/03/GettyImages-139720130.jpg",
-        position: "Workshop Chair",
-        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
-        fun_fact: "",
-        email: "bfoey@ucsd.edu",
-        github: "https://github.com/BrandonFoey"
-    },
-    {
-        name: "Bao Hoang",
-        image: "https://cdn1-www.dogtime.com/assets/uploads/2019/04/DogSilly.jpg",
-        position: "Event Chair",
-        link: "https://github.com/cqvu/os-ucsd.ucsd.edu",
-        fun_fact: '"I have a good personality"',
-        email: "bghoang@ucsd.edu",
-        github: "https://github.com/bghoang"
-    }
-];
+import officers from '../data/officer-data';
+
+import '../css/about.css';
 
 class About extends React.Component {
     constructor(props) {
@@ -85,7 +38,7 @@ class About extends React.Component {
 
                 {/* MEET THE BOARD */}
                 <h2 style={{ fontSize: "40px", textAlign: "center", padding: "10px" }}>Meet the Board</h2>
-                <Grid container spacing={1.5}>
+                <div className="officer-outer-container">
                     {officers.map(officer => (
                         <Officer
                             name={officer.name}
@@ -97,13 +50,42 @@ class About extends React.Component {
                             github={officer.github}
                         />
                     ))}
-                </Grid>
+                </div>
                 <Footer />
             </div >
         )
     }
 }
 
+const Officer = props => {
+    const {
+        name, image, position, fun_fact, link,
+        email, github
+    } = props;
+
+    return (
+        <div className="officer-container">
+            <a href={link}>
+                <img src={image} className="officer-img"/>
+            </a>
+            <div className="officer-info">
+                <h2>{name}</h2>
+                <h4>{position}</h4>
+                <p>{fun_fact}</p>
+                <div className="officer-contact">
+                    <a className="officer-link-icon" href={"mailto:" + email}>
+                        <EmailIcon style={{ fill: "#314858", width: 25, height: 25 }} />
+                    </a>
+                    <a className="officer-link-icon" href={github}>
+                        <GitHubIcon style={{ fill: "#314858", width: 25, height: 25 }} />
+                    </a>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+/*
 const Officer = props => {
     function over(e) {
         e.target.width = 280;
@@ -155,5 +137,6 @@ const Officer = props => {
         </div>
     );
 };
+*/
 
 export default About
